Move geolocatino to DBW
